+++ NEWS.md
@@ -1,3 +1,8 @@
+# rAmCharts4 1.1.0.9000 (2020-11-xx)
+
+* New option `colors` in `amStackedBarChart`, to set the colors of the bars.
+
+
 # rAmCharts4 1.1.0 (2020-10-30)
 
 * Updated 'amcharts4'.